refactor reporting portfolio review route
